feat(blacklist): re-arrange some blacklist structure, rename clear to delete
