+++ src/boorubot/blacklist/index.ts
@@ -1,5 +1,5 @@
 import { SleetSlashCommand, SleetSlashCommandGroup } from 'sleetcord'
-import { blacklistClear } from './clear.js'
+import { blacklistDelete } from './delete.js'
 import { blacklistList } from './list.js'
 import { blacklistAddSite, blacklistRemoveSite } from './sites.js'
 import { blacklistAddTags, blacklistRemoveTags } from './tag.js'
@@ -19,6 +19,6 @@ export const blacklistRemove = new SleetSlashCommandGroup({
 export const blacklist = new SleetSlashCommand({
   name: 'blacklist',
   description: 'Blacklist tags or sites from being returned by the bot',
-  options: [blacklistAdd, blacklistRemove, blacklistList, blacklistClear],
+  options: [blacklistAdd, blacklistRemove, blacklistList, blacklistDelete],
   default_member_permissions: ['ManageMessages'],
 })
