+++ src/boorubot/utils.ts
@@ -1,5 +1,5 @@
 import { BooruConfig } from '@prisma/client'
-import { bold, CommandInteraction, escapeCodeBlock } from 'discord.js'
+import { CommandInteraction } from 'discord.js'
 import { database } from '../util/db.js'
 
 export function getReferenceIdFor(interaction: CommandInteraction): string {
@@ -24,77 +24,6 @@ export async function ensureConfigFor(
       referenceId,
     },
   })
-}
-
-export interface Blacklist {
-  tags: string[]
-  sites: string[]
-}
-
-export interface FormatBlacklistOptions {
-  highlightTags?: string[]
-  highlightSites?: string[]
-}
-
-export function formatBlacklist(
-  blacklist: Blacklist,
-  { highlightTags = [], highlightSites = [] }: FormatBlacklistOptions = {},
-): string {
-  const tags = formatBlacklistArray(blacklist.tags, highlightTags)
-  const sites = formatBlacklistArray(blacklist.sites, highlightSites)
-
-  return `\`\`\`asciidoc
-Tags :
-======
-${tags}
-
-Sites:
-======
-${sites}
-\`\`\``
-}
-
-function formatBlacklistArray(items: string[], highlight: string[]): string {
-  if (items.length === 0) {
-    return '[Empty]'
-  }
-
-  return items
-    .map((item) => escapeCodeBlock(item))
-    .map((item) => (highlight.includes(item) ? bold(item) : item))
-    .join(', ')
-}
-
-export function getBlacklistFor(referenceId: string): Promise<Blacklist> {
-  return database.booruConfig
-    .findFirst({
-      where: {
-        referenceId,
-      },
-      select: {
-        tags: {
-          select: {
-            name: true,
-          },
-        },
-        sites: {
-          select: {
-            name: true,
-          },
-        },
-      },
-    })
-    .then((result) =>
-      result
-        ? {
-            tags: result.tags.map((tag) => tag.name),
-            sites: result.sites.map((site) => site.name),
-          }
-        : {
-            tags: [],
-            sites: [],
-          },
-    )
 }
 
 /**
